+++ src/main/java/org/cardanofoundation/authentication/repository/WalletRepository.java
@@ -4,6 +4,7 @@ import org.cardanofoundation.authentication.model.entity.WalletEntity;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -18,5 +19,16 @@ public interface WalletRepository extends JpaRepository<WalletEntity, Long> {
   @Query(value = "SELECT we FROM UserEntity ue "
       + "JOIN WalletEntity we ON ue.id = we.user.id "
       + "WHERE we.address = :address")
-  Optional<WalletEntity> findWalletByAddress(String address);
+  Optional<WalletEntity> findWalletByAddress(@Param("address") String address);
+
+  /*
+   * @author: phuc.nguyen5
+   * @since: 19/05/2023
+   * description: find address by userId
+   * @update:
+   */
+  @Query(value = "SELECT we.address FROM UserEntity ue "
+      + "JOIN WalletEntity we ON ue.id = we.user.id "
+      + "WHERE ue.id = :userId")
+  String findAddressByUserId(@Param("userId") Long userId);
 }