+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## [0.1.2](https://github.com/cardano-foundation/cf-explorer-authentication/compare/v0.1.1...v0.1.2) (2023-05-31)
+
+
+### Features
+
+* add email + stake key for user ([1f58c60](https://github.com/cardano-foundation/cf-explorer-authentication/commit/1f58c607bac00cdd63851005f4932c52349fb247))
+* refactor code after remove username ([884cb06](https://github.com/cardano-foundation/cf-explorer-authentication/commit/884cb0620f9ab80ee9c3a79477762a209374f38e))
+* remove username when sign up ([25310e4](https://github.com/cardano-foundation/cf-explorer-authentication/commit/25310e4bbb771fd5f5cb977199b12118a28fba37))
+* set address for user info ([abed61f](https://github.com/cardano-foundation/cf-explorer-authentication/commit/abed61f98bf8a65c2e8413066e79e1dab79e531a))
+
+
+### Bug Fixes
+
+* handle bug relate upgrade version ([a3ae556](https://github.com/cardano-foundation/cf-explorer-authentication/commit/a3ae5568ae62163b7ac48780b8768b53953d0026))
+
+## [0.1.1](https://github.com/cardano-foundation/cf-explorer-authentication/compare/v0.1.0...v0.1.1) (2023-05-15)
+
+
+### Bug Fixes
+
+* Update README.md trigger release ([c0f092e](https://github.com/cardano-foundation/cf-explorer-authentication/commit/c0f092e48d3218e7cc36dfbfdd17e7612003156b))
+
 ## 0.1.0 (2023-05-15)